rename wpath to ppath; it is now pages, not webcore
