+++ pages/ppath/ppath.go
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// Package wpath handles pages paths.
-package wpath
+// Package ppath handles pages paths.
+package ppath
 
 import (
 	"path"
